+++ packages/manager/src/features/Databases/DatabaseDetail/DatabaseSettings/DatabaseSettingsResetPasswordDialog.tsx
@@ -5,7 +5,7 @@ import { ActionsPanel } from 'src/components/ActionsPanel/ActionsPanel';
 import { ConfirmationDialog } from 'src/components/ConfirmationDialog/ConfirmationDialog';
 import { Notice } from 'src/components/Notice/Notice';
 import { Typography } from 'src/components/Typography';
-import { useDatabaseCredentialsMutation } from 'src/queries/databases';
+import { useDatabaseCredentialsMutation } from 'src/queries/databases/databases';
 
 interface Props {
   databaseEngine: Engine;
