+++ packages/manager/src/hooks/useEventHandlers.ts
@@ -1,7 +1,7 @@
 import { useQueryClient } from '@tanstack/react-query';
 
 import { oauthClientsEventHandler } from 'src/queries/account/oauth';
-import { databaseEventsHandler } from 'src/queries/databases';
+import { databaseEventsHandler } from 'src/queries/databases/events';
 import { domainEventsHandler } from 'src/queries/domains';
 import { firewallEventsHandler } from 'src/queries/firewalls';
 import { imageEventsHandler } from 'src/queries/images';
